Make it possible to run tests on windows (with a little cygwin help)
