+++ testdb/supervisord-cygwin.conf
@@ -1,0 +1,61 @@
+[supervisord]
+logfile = %(here)s/supervisord.log
+pidfile = %(here)s/supervisord.pid
+#nodaemon = true
+
+[inet_http_server]
+port = 127.0.0.1:9001
+
+[supervisorctl]
+serverurl = http://127.0.0.1:9001
+
+[rpcinterface:supervisor]
+supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
+
+[program:db1]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --dbpath %(ENV_DIRECTORY)sdb1 --bind_ip=127.0.0.1 --port 40001
+
+[program:db2]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --dbpath %(ENV_DIRECTORY)sdb2 --bind_ip=127.0.0.1 --port 40002 --auth 
+
+[program:rs1a]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs1 --dbpath %(ENV_DIRECTORY)srs1a --bind_ip=127.0.0.1 --port 40011
+[program:rs1b]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs1 --dbpath %(ENV_DIRECTORY)srs1b --bind_ip=127.0.0.1 --port 40012
+[program:rs1c]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs1 --dbpath %(ENV_DIRECTORY)srs1c --bind_ip=127.0.0.1 --port 40013
+ 
+[program:rs2a]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs2 --dbpath %(ENV_DIRECTORY)srs2a --bind_ip=127.0.0.1 --port 40021
+[program:rs2b]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs2 --dbpath %(ENV_DIRECTORY)srs2b --bind_ip=127.0.0.1 --port 40022
+[program:rs2c]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs2 --dbpath %(ENV_DIRECTORY)srs2c --bind_ip=127.0.0.1 --port 40023
+
+[program:rs3a]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs3 --dbpath %(ENV_DIRECTORY)srs3a --bind_ip=127.0.0.1 --port 40031 --auth --keyFile=%(ENV_DIRECTORY)skeyfile
+[program:rs3b]                                                                                                                                                                                                      
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs3 --dbpath %(ENV_DIRECTORY)srs3b --bind_ip=127.0.0.1 --port 40032 --auth --keyFile=%(ENV_DIRECTORY)skeyfile
+[program:rs3c]                                                                                                                                                                                                      
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs3 --dbpath %(ENV_DIRECTORY)srs3c --bind_ip=127.0.0.1 --port 40033 --auth --keyFile=%(ENV_DIRECTORY)skeyfile
+
+[program:rs4a]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --shardsvr --replSet rs4 --dbpath %(ENV_DIRECTORY)srs4a --bind_ip=127.0.0.1 --port 40041
+
+[program:cfg1]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --configsvr --dbpath %(ENV_DIRECTORY)scfg1 --bind_ip=127.0.0.1 --port 40101
+
+[program:cfg2]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --configsvr --dbpath %(ENV_DIRECTORY)scfg2 --bind_ip=127.0.0.1 --port 40102
+
+[program:cfg3]
+command = mongod --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --configsvr --dbpath %(ENV_DIRECTORY)scfg3 --bind_ip=127.0.0.1 --port 40103 --auth --keyFile=%(ENV_DIRECTORY)skeyfile
+
+[program:s1]
+command = mongos --configdb 127.0.0.1:40101 --bind_ip=127.0.0.1 --port 40201 --chunkSize 1
+
+[program:s2]
+command = mongos --configdb 127.0.0.1:40102 --bind_ip=127.0.0.1 --port 40202 --chunkSize 1
+
+[program:s3]
+command = mongos --configdb 127.0.0.1:40103 --bind_ip=127.0.0.1 --port 40203 --chunkSize 1 --keyFile=%(ENV_DIRECTORY)skeyfile
